refactor: add header and details
