+++ src/components/details/details.interfaces.ts
@@ -1,0 +1,6 @@
+import { ApiResult } from 'api/api.interfaces';
+
+export interface DetailsProps {
+  details: ApiResult;
+  closeDetailsHandler: () => void;
+}
