+++ src/components/header/header.interfaces.ts
@@ -1,0 +1,4 @@
+export interface HeaderProps {
+  page: number;
+  onSearchHandler: (currentPage: number, savedValue?: string | null) => Promise<void>;
+}
