+++ src/components/header/header.module.scss
@@ -1,0 +1,13 @@
+.header {
+  display: flex;
+  gap: 10px;
+  justify-content: center;
+}
+
+.searchButton {
+  background-color: limegreen;
+}
+
+.crashButton {
+  background-color: crimson;
+}
