+++ src/components/pagination/Pagination.tsx
@@ -1,0 +1,23 @@
+import { RESULTS_COUNT_ON_PAGE } from 'api/api.consts';
+import { FC, memo } from 'react';
+
+import { PaginationProps } from './pagination.interfaces';
+import styles from './pagination.module.scss';
+
+export const Pagination: FC<PaginationProps> = memo(({ page, count, setPageHandler }) => {
+  const pages = new Array(Math.ceil(count / RESULTS_COUNT_ON_PAGE)).fill(null);
+
+  return (
+    <div className={styles.pagination}>
+      {pages.map((_, index) => (
+        <div
+          key={index}
+          className={styles.page + ' ' + (page === index + 1 ? styles.active : '')}
+          onClick={() => setPageHandler(index + 1)}
+        >
+          {index + 1}
+        </div>
+      ))}
+    </div>
+  );
+});
