+++ src/components/pagination/pagination.interfaces.ts
@@ -1,0 +1,5 @@
+export interface PaginationProps {
+  page: number;
+  count: number;
+  setPageHandler: (pageNumber: number) => void;
+}
