+++ src/components/pagination/pagination.module.scss
@@ -1,0 +1,28 @@
+.pagination {
+  display: flex;
+  gap: 10px;
+  justify-content: center;
+}
+
+.page {
+  cursor: pointer;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 40px;
+  height: 40px;
+
+  border-radius: 5px;
+  box-shadow: 0 0 5px black;
+
+  transition:
+    color 0.2s,
+    background-color 0.2s;
+
+  &.active {
+    color: white;
+    background-color: black;
+  }
+}
