+++ src/components/results/results.interfaces.ts
@@ -4,9 +4,6 @@ import { Dispatch, SetStateAction } from 'react';
 
 export interface ResultsProps {
   results: ApiResults;
-  count: number;
-  page: number;
-  setPageHandler: (pageNumber: number) => void;
   setDetails: Dispatch<SetStateAction<ApiResult | null>>;
   setIsLoading: Dispatch<SetStateAction<boolean>>;
 }
