+++ src/components/results/results.module.scss
@@ -1,14 +1,3 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  align-items: center;
-  justify-content: space-between;
-
-  height: 100%;
-  padding: 20px;
-}
-
 .results {
   display: flex;
   flex-direction: column;
@@ -33,34 +22,5 @@
   &::first-letter {
     font-size: 1.3em;
     font-size: bold;
-  }
-}
-
-.pagination {
-  display: flex;
-  gap: 10px;
-  justify-content: center;
-}
-
-.page {
-  cursor: pointer;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  width: 40px;
-  height: 40px;
-
-  border-radius: 5px;
-  box-shadow: 0 0 5px black;
-
-  transition:
-    color 0.2s,
-    background-color 0.2s;
-
-  &.active {
-    color: white;
-    background-color: black;
   }
 }
